[Aardvark.Base] added additional PixImage constructors

+++ src/Aardvark.Base/Math/Color.cs
@@ -413,31 +413,31 @@ namespace Aardvark.Base
             return -1;
         }
 
-        private static Dict<Tup<Type, int>, Format> s_defaultFormatMap =
-            new Dict<Tup<Type, int>, Format>
+        private static Dict<Tup<Type, long>, Format> s_defaultFormatMap =
+            new Dict<Tup<Type, long>, Format>
             {
-                { Tup.Create(typeof(byte), 1), Format.Gray },
-                { Tup.Create(typeof(byte), 3), Format.BGR },
-                { Tup.Create(typeof(byte), 4), Format.BGRA },
+                { Tup.Create(typeof(byte), 1L), Format.Gray },
+                { Tup.Create(typeof(byte), 3L), Format.BGR },
+                { Tup.Create(typeof(byte), 4L), Format.BGRA },
 
-                { Tup.Create(typeof(ushort), 1), Format.Gray },
-                { Tup.Create(typeof(ushort), 3), Format.RGB },
-                { Tup.Create(typeof(ushort), 4), Format.RGBA },
+                { Tup.Create(typeof(ushort), 1L), Format.Gray },
+                { Tup.Create(typeof(ushort), 3L), Format.RGB },
+                { Tup.Create(typeof(ushort), 4L), Format.RGBA },
 
-                { Tup.Create(typeof(uint), 1), Format.Gray },
-                { Tup.Create(typeof(uint), 3), Format.RGB },
-                { Tup.Create(typeof(uint), 4), Format.RGBA },
+                { Tup.Create(typeof(uint), 1L), Format.Gray },
+                { Tup.Create(typeof(uint), 3L), Format.RGB },
+                { Tup.Create(typeof(uint), 4L), Format.RGBA },
 
-                { Tup.Create(typeof(float), 1), Format.Gray },
-                { Tup.Create(typeof(float), 3), Format.RGB },
-                { Tup.Create(typeof(float), 4), Format.RGBA },
+                { Tup.Create(typeof(float), 1L), Format.Gray },
+                { Tup.Create(typeof(float), 3L), Format.RGB },
+                { Tup.Create(typeof(float), 4L), Format.RGBA },
 
-                { Tup.Create(typeof(double), 1), Format.Gray },
-                { Tup.Create(typeof(double), 3), Format.RGB },
-                { Tup.Create(typeof(double), 4), Format.RGBA },
+                { Tup.Create(typeof(double), 1L), Format.Gray },
+                { Tup.Create(typeof(double), 3L), Format.RGB },
+                { Tup.Create(typeof(double), 4L), Format.RGBA },
             };
 
-        public static Format FormatDefaultOf(this Type type, int channelCount)
+        public static Format FormatDefaultOf(this Type type, long channelCount)
         {
             Format format;
             if (s_defaultFormatMap.TryGetValue(Tup.Create(type, channelCount),
